Move App.js to component folder
